Add files via upload
I just vib coded the "file-rename.py" script using chat gbt and it works. I always forget to change 3d model names when I download them and sites like makerworld.com add "+" to the file names and sub-directories, including when zip is downloaded. This is ugly and I like the "-" in filenames, not "+".